Asm: rename `@symbols` to `@import` and do lazy loading
I realised that, whenever using `@symbols`, more often than not you
want to also load the labels' content into the linker (needing semantic
analysis on the entire file). Otherwise, the linker will complain about
an undefined symbol when referencing a label and not manually specifying
the file.

It's also just annoying to always have to list the files. Like Zig, only
the root file is needed. If wanted, you can specify others but they'll
be pruned by dead-code-elimination if that's enabled. Or maybe later
more (optional) 'root sections' are added.

If only @defines are needed, and not the content (like userspace using
constants from a kernel lib), the @defines should be separated from
opaques.